feat: add objectFilter to page
